% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Bubble Plot with Choice on Color Size and Text}
\usage{
bubble_plot(dataframe,ctg.idx,num.idx,size.idx,color.idx,
                   text.idx,colors,xaxis_name,xaxis_format,yaxis_name,
                   yaxis_format,legend_pos_y=1,title,...)
}
\arguments{
\item{...}{other parameters for plotting. See \code{\link[plotly]{layout}}
for more information}

\item{dataframe:}{a dataframe object}

\item{ctg.idx:}{a character or an integer, indicate which column will be selected as the
variable to show in x axis}

\item{num.idx:}{a character or an integer, indicate which column will be selected as the
variable to show in y axis}

\item{size.idx:}{a character or an integer, indicate which column will be selected as the
variable to be assigned to the bubble's size, should be a numeric column}

\item{color.idx:}{a character or and integer, indicate which column will be selected as the
variable to be assigned to the bubble's color, perfer categorical column}

\item{text.idx:}{a character or an integer, indicate which column will be selected as the
variable shown in the hover label}

\item{colors:}{a vector, the colors that to be used in making bubbles}

\item{xaxis_name:}{name of x axis}

\item{xaxis_format:}{format of x axis. See \code{\link[plotly]{layout}}}

\item{yaxis_name:}{name of y axis}

\item{yaxis_format:}{format of y axis. See \code{\link[plotly]{layout}}}

\item{legend_pos_y:}{a numeric, indicate the postion of the title of legend on vertical
axis, default to 1}

\item{title:}{the title of the plot}
}
\value{
This function will return an object of class plotly. By calling the function you
will get a bubble plot where X axis value, Y axis value, color, size and text is your choice.
}
\description{
\code{bubble_plot} makes bubble plot where each bubble represent a sample, with customer
choiced X axis, Y axis, size, color and text
}
\details{
The column of size must be a numerical column, otherwise the mapping of size will generate
error.
}
\examples{
data("machinery_fin_charts")
test_df=machinery_fin_charts\%>\%
  filter(年份==2017 & 分类=='通用设备')
bubble_plot(dataframe=test_df,ctg.idx=15,num.idx='毛利率',size.idx='主营业务收入',
            color.idx='经济区划分',text.idx=7,colors=brewer.pal(8,'Set1'),
            xaxis_name='营收CAGR5',yaxis_name='毛利率',xaxis_format='\%',yaxis_format='\%',
            title='通用设备企业实力气泡图（2017）',paper_bgcolor='#ccece6')
test_df=machinery_fin_charts\%>\%
  filter(年份==2017 & 分类=='电力设备')
bubble_plot(dataframe=test_df,ctg.idx=10,num.idx='毛利率',size.idx='主营业务收入',
            color.idx=15,text.idx=7,
            colors='Reds',xaxis_name='员工总数（人）',yaxis_name='毛利率',
            xaxis_format='',yaxis_format='\%',legend_pos_y = 1.02,
            title='电力设备企业实力气泡图（2017）',paper_bgcolor='#ccece6')
}
\seealso{
\code{\link[plotly]{plot_ly}},\code{\link[plotly]{layout}}
}
