% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pct.R
\name{cal_pct}
\alias{cal_pct}
\title{Calculate Percentage of Each Category}
\usage{
cal_pct(dataframe,ctg.idx,num.idx,condition,condition.idx)
}
\arguments{
\item{dataframe:}{a dataframe object}

\item{ctg.idx:}{a character or an integer, indicate which column will be selected as the
categorical variable}

\item{num.idx:}{a character or an integer, indicate which column will be selected as the
numerical variable}

\item{condition:}{a character or a numeric, indicate whether to select part of the dataframe}

\item{condition.idx:}{a character or an integer, indicate which column is the condition in.
Should be given some value only if the condition appears in multiple columns in the dataframe}
}
\value{
A new dataframe contains three columns. The first is a categorical variable,
the second is a numerical variable, and the third is the percentage of each level in
the categorical variable with respect to the numerical variable.
}
\description{
\code{cal_pct} calculate the percentage of each category, and can be based on some criteria
about selecting data, like the market share of each company in a certain year.
}
\details{
This function uses dplyr package which offers a fast computation of percentage. It can take
either column name or column index as the input to specify the categorical and numerical
variable, as well as using some condition to subset data. The function can detect the
column in which the condition is in automatically, so you are not required to input it.
}
\examples{
data("dairy_fin_charts")
cal_pct(dairy_fin_charts,ctg.idx = 'name',num.idx = "income",condition=2017)
## another way to get the same result
cal_pct(dairy_fin_charts[dairy_fin_charts$Year==2017,],ctg.idx = 1,num.idx = 2)
cal_pct(dairy_fin_charts,ctg.idx = 1,num.idx = 2,condition=2010)
}
