% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_class.R
\name{column_class}
\alias{column_class}
\title{Separate features by categorical or numerical}
\usage{
column_class(dataframe)
}
\arguments{
\item{dataframe:}{a dataframe object}
}
\value{
A list with two element, both of them are dataframe, which stores the
column name and column index of numerical and categorical variables.
}
\description{
\code{column_class} get the class information for each column vector of a dataframe.
}
\details{
This function receives a dataframe and returns a list which contains two dataframe:
numerical and categorical, which contain the column name and location
of numerical and categorical columns of the input dataframe, respectively.
}
\examples{
machinery_fin_charts_class=column_class(machinery_fin_charts)
machinery_fin_charts_class$numerical
machinery_fin_charts_class$categorical
}
