% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_check.R
\name{corr_check}
\alias{corr_check}
\title{Correlation Checking Plot}
\usage{
corr_check(dataframe,eliminate)
}
\arguments{
\item{dataframe:}{a dataframe object}

\item{eliminate:}{a vector of numeric or character, indicate the column names or column
index of those columns that you do not want to include in correlation plot. Usually id}
}
\value{
A correlation plot with correlation coefficients shown in the upper panel and
points and linear regression line plot shown in the lower panel, for choosen
numerical columns.
}
\description{
\code{corr_check} gives a nice plot for checking correlation between numerical variables.
}
\details{
The upper panel shows the correlation coefficients of numerical columns. The lower
panel shows the points and regression line of corresponding numerical columns, where
one can get some basic sense about how those variables are correlated.
The color displayed is an easy way to know the correlation between variables.
If the color is dark blue, then their are strong negative correlation (less than -0.6),
if it is light blue, then some negative correlation exist (-0.4 to -0.6),
if it is green, then you may think the two variables have little correlation (-0.4 to 0.4)
if it is pink, then there is relatively strong positive correlation (0.4 to 0.6)
and if it is red, then they have string positive correlation (bigger than 0.6).
}
\examples{
data("macro_data_chn")
corr_check(dataframe=macro_data_chn,eliminate = 'year')
## another example, add a categorical variable,
## see if the function eliminate it automatically
test_df=macro_data_chn
test_df$ctg='cat'
corr_check(dataframe = test_df,eliminate=c(1,2))
}
\seealso{
\code{\link[graphics]{pairs}}
}
