% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dairy_fin_charts.R
\docType{data}
\name{dairy_fin_charts}
\alias{dairy_fin_charts}
\title{Financial Charts of Top Chinese Dairy Industry Companies(2006-2017)}
\format{A dataframe containing 120 rows and 5 columns}
\usage{
data(dairy_fin_charts)
}
\description{
Income, profit and ROE of ten Chinese dairy companies trade on stock market,
from 2006 to 2017.
}
\references{
Wind database of Chinese and Hong Kong's stock market data
}
\keyword{datasets}
