% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_axis.R
\name{double_axis}
\alias{double_axis}
\title{Stack Bar Plot and Line Plot with Double Axises}
\usage{
double_axis(dataframe,ctg.idx,lines.idx,bars.idx,condition,condition.idx,
                   lines.mode,lines.colors,lines.width=2,lines.names,
                   bars.colors,bars.names,xaxis_name,line.axis_format="",
                   line.axis_name,bar.axis_name,title,annOn,...)
}
\arguments{
\item{...}{other parameters for plotting, see "Details" and \code{\link[plotly]{layout}}
for more information}

\item{dataframe:}{a dataframe object}

\item{ctg.idx:}{a character or an integer, indicate which column will be selected as the
variable to show in x axis}

\item{lines.idx:}{a vector of characters or integers, but not both, indicate which
column(s) will be selected to make line plot}

\item{bars.idx:}{a vector of characters or integers, but not both, indicate which
column(s) will be selected to make bar plot}

\item{condition:}{a vector, indicate how to select part of the dataframe}

\item{condition.idx:}{a character or an integer, indicate which column is the condition in.
Should be given some value if the condition appears in multiple columns in the dataframe}

\item{lines.mode:}{a string indicating the mode of line plot, can be 'lines' or 'lines+markers'}

\item{lines.colors:}{a vector, the palette for line plot}

\item{lines.width:}{a numeric, the width of the line plot, default to 2}

\item{lines.names:}{a vector of strings, indicating the name of line variables in legend}

\item{bars.colors:}{a vector, the palette for bar plot}

\item{bars.names:}{a vector of strings, indicating the name of bar variables in legend}

\item{xaxis_name:}{a string, the name of x axis}

\item{line.axis_format:}{a string, the format of y axis of line plot.
See \code{\link[plotly]{layout}} for more information, default to ""}

\item{line.axis_name:}{a string, the name of y axis for line plot, will be displayed
on the left side of the plot}

\item{bars.axis_name:}{a string, the name of y axis for bar plot, will be displayed
on the right side of the plot}

\item{title:}{title of the plot}

\item{annOn:}{a logical, decide whether to add annotations for line plot.
See "Details" for more information}
}
\value{
This function will return an object of class plotly. By calling the function you
will plot with double y axis, one for display lines and the other for the bars.
}
\description{
\code{double_axis} makes plot with two y axises, one for display line plot, and the other
for bar plot
}
\details{
This function is useful when the numerical variable you want to display have
various range, like some are percentage while others are values. The three dots
structure makes it easy for advanced users to set other graphing parameters, such
as legend, paper_bgcolor and margin. This function can be combined with
\code{\link[plotly]{add_trace}} to add more traces, such as points, to the plot.
Note that one should carefully choose how many variables to display and
whether to add annotations to make the plot clear and fancy.
}
\examples{
## a warm up example
data("dairy_fin_charts")
test_df=dairy_fin_charts\%>\%
  filter(name=='伊利股份')
double_axis(dataframe=test_df,ctg.idx='Year',lines.idx='profit',bars.idx=2,
            lines.mode='lines+markers',lines.colors='rgb(128,0,128)',
            lines.names = '营业利润',bars.colors = 'rgba(55,128,192,0.7)',
            bars.names = '营业收入',xaxis_name = '年份',
            line.axis_name = '营业利润（亿元）',bar.axis_name = '营业收入（亿元）',
            title='伊利股份营收及利润',annOn=T,margin=list(r=40))
## a more complex example
data("estate_fin_charts")
double_axis(dataframe=estate_fin_charts,ctg.idx='Year',lines.idx=c(5,6),
            bars.idx=c('asset','liability'),condition='万科A',lines.mode = 'lines',
            lines.colors = c("rgb(128, 0, 128)",'rgb(255,140,0)'),
            lines.width = 4,lines.names=c('总资产收益率','净资产收益率'),
            bars.colors = c('rgba(55,128,192,0.7)','rgba(219, 64, 82,0.7)'),
            bars.names = c('总资产','总负债'),xaxis_name = '年份',
            line.axis_format = '',line.axis_name='百分比',bar.axis_name = '单位：亿元',
            title='财务分析（万科A）',annOn=F,
            legend=list(x=0.45,y=1.03,orientation='h',
                        font=list(size=10),bgcolor="transparent"),
            margin=list(r=54),paper_bgcolor='#ccece6')
\dontrun{
## be careful when set annOn when have more than one lines
## the plot may not be clear
double_axis(dataframe=estate_fin_charts,ctg.idx='Year',lines.idx=c(5,6),
            bars.idx=c('asset','liability'),condition='万科A',lines.mode = 'lines+markers',
            lines.colors = c("rgb(128, 0, 128)",'rgb(255,140,0)'),
            lines.width = 4,lines.names=c('总资产收益率','净资产收益率'),
            bars.colors = c('rgba(55,128,192,0.7)','rgba(219, 64, 82,0.7)'),
            bars.names = c('总资产','总负债'),xaxis_name = '年份',
            line.axis_format = '',line.axis_name='百分比',bar.axis_name = '单位：亿元',
            title='财务分析（万科A）',annOn=T,
            legend=list(x=0.45,y=1.03,orientation='h',
                        font=list(size=10),bgcolor="transparent"),
            margin=list(r=54),paper_bgcolor='#ccece6')
}
}
\seealso{
\code{\link[plotly]{add_trace}},\code{\link[plotly]{layout}}
}
\author{
Jizhou Kang <jizhou_kang@hotmail.com>
}
