% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dppbar_imports.R
\docType{package}
\name{dppbar-package}
\alias{dppbar}
\alias{dppbar-package}
\title{dppbar: Data Processing and Plotting for Business Analysis in R}
\description{
A toolbox of data processing and plotting methods in R. It combines multiple data cleaning, manipultaing and statistical summarying methods together with fancy visualization tools simplified from other packages to help people with little R programming experience to accomplish exploratory data analysis for their own goal in R fastly. The package has a designed emphasis on financial and economic data, especially balance sheet and panel data.
}
\keyword{internal}
