% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estate_fin_charts.R
\docType{data}
\name{estate_fin_charts}
\alias{estate_fin_charts}
\title{Financial Charts of Chinese Real Estate Companies(2007-2016)}
\format{A dataframe containing 990 rows and 36 columns}
\usage{
data(estate_fin_charts)
}
\description{
Balance sheet and characteristic data of Chinese real estate companies whose stock is traded on
Chinese and Hong Kong's stock market together with macroeconomic data of China,
such as GDP growth rate and interest rate, from 2007 to 2016.
}
\references{
Wind database of Chinese and Hong Kong's stock market data and
Chinese industry and macroeconomic data
}
\keyword{datasets}
