% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_bar.R
\name{facet_bar}
\alias{facet_bar}
\title{Bar Plots in Small Facet}
\usage{
facet_bar(dataframe,ctg.idx,num.idx,condition.idx,label.idx,
                 legend_name,legend_label,colors,
                 xaxis_name,xaxis_label,yaxis_name,title,
                 type='histogram',stack=F,paper_bgcolor="#f2f2f2")
}
\arguments{
\item{dataframe:}{a dataframe object}

\item{ctg.idx:}{a character or an integer, indicate which column will be selected as the
categorical variable shown in x axis}

\item{num.idx:}{a vector of character or an integer, indicate which column
will be selected as the numerical variable shown in y axis,
only valid when "type" is bar}

\item{condition.idx:}{a character or an integer, indicate which column will be treated as
the legend}

\item{label.idx:}{a character or an integer, indicate to which column should the split of
facets been done}

\item{legend_name:}{title of the legend}

\item{legend_label:}{a vector of string, name of legend ticks}

\item{colors:}{a vector or a character, the palette used for plotting}

\item{xaxis_name:}{name of x axis}

\item{xaxis_label:}{name of ticks of the x axis}

\item{yaxis_name:}{name of y axis}

\item{title:}{name of the plot}

\item{type:}{a character of either 'histogram' or 'bar', indicating the way for drawing
bars, default to 'histogram'. See "Details" for more information}

\item{stack:}{a logical indicate whether the bar should be stack or dodged,
default to TURE which means stack}

\item{paper_bgcolor:}{background color of the whole plot, default to "#f2f2f2"}
}
\value{
This function will return an object of class ggplot. By calling the function you
will get a plot with multiple facet where each of them displays a bar chart or histogram.
}
\description{
\code{facet_bar} separates a whole bar plot into smaller bar plots displayed in small
facets. It is used to show multiple (up to three) categorical information. A similar
plot can be found in package \code{"lattice"} where it is called \code{\link{barchart}}
}
\details{
If "num.idx" is missing, then only histogram can be plot, where the height of each bar
equals to the count of that specific category, selected by "ctg.idx", "condition.idx",
and "label.idx". On the other hand, if a num.idx is choosen, then you need to switch "type"
to "bar" where the total of the numerical variables you selected that belong to a particular
category will be used as the height of bar.
}
\examples{
data("tmall_milk_sales")
facet_bar(dataframe=tmall_milk_sales,ctg.idx='label',num.idx=9,
          condition.idx='promotion',label.idx='brand',legend_name='是否促销',
          legend_label=c('不促销','促销'),
          colors=c('#A6CEE3','#1F78B4'),xaxis_name='产品类型',
          xaxis_label=c('热销产品','普通产品','滞销产品'),
          yaxis_name='价格（元）',title='线上乳制品分类销售情况',
          type='bar',stack=T,paper_bgcolor='#ccece6')
## another type of plot
facet_bar(dataframe=tmall_milk_sales,ctg.idx='class',num.idx=NA,
          condition.idx='label',label.idx='brand',legend_name='产品类型',
          legend_label=c('热销产品','普通产品','滞销产品'),
          colors=brewer.pal(3,'Set2'),xaxis_name='分类',xaxis_label=c('酸奶','牛奶'),
          yaxis_name='产品数（件）',title='线上乳制品分类销售情况',
          paper_bgcolor='#ccece6')
}
\seealso{
\code{\link[ggplot2]{facet_wrap}}
}
\author{
Jizhou Kang <jizhou_kang@hotmail.com>
}
