% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizontal_bar.R
\name{horizontal_bar}
\alias{horizontal_bar}
\title{Horizontal Bar Plot Showing Percentage}
\usage{
horizontal_bar(dataframe,h.idx,v.idx,h_name,v_name,colors,
                      xaxis_name,title,...)
}
\arguments{
\item{...}{other parameters for plotting, see "Details" and \code{\link[plotly]{layout}}
for more information}

\item{dataframe:}{a dataframe object}

\item{h.idx:}{a character or an integer, indicate which column will be selected as the
variable to show in horizontal axis, should be a categorical variable}

\item{v.idx:}{a character or an integer, indicate which column will be selected as the
variable to show in vertical axis, should be a categorical variable}

\item{h_name:}{a vector, indicate the label shows in horizontal axis}

\item{v_name:}{a vector, indicate the label shows in vertical axis}

\item{colors:}{a vector, the palette used to plot the bar}

\item{xaxis_name:}{a character, the name of x axis}

\item{title:}{a character, the name of the plot}
}
\value{
This function will return an object of class plotly. By calling the function you
will get a horizontal bar plot where each bar shows the percentage of samples belong to
a category with respect to another cateogry.
}
\description{
\code{horizontal_bar} plot a horizontal bar plot, with each small bar represent the
percentage of a categorical variable with respect to another categorical variable.
}
\details{
The function will first count the number of samples based on the two categorical variables
specified by 'h.idx' and 'v.idx', then calculate the percentage of 'h.idx' variable
to 'v.idx' variable. When do plotting, one should take good care of the order of 'v_name'
and 'h_name'.
}
\examples{
data("tmall_milk_sales")
horizontal_bar(dataframe=tmall_milk_sales,h.idx='label',v.idx='pack',
               h_name=c('热销产品','普通产品','滞销产品'),
               v_name=c('爱克林包装','杯装','袋装','盒装','瓶装'),
               colors = brewer.pal(3,'Set1'),xaxis_name='百分比',
               title='电商乳制品产品线分类统计')
horizontal_bar(dataframe=tmall_milk_sales,h.idx='feature',v.idx='label',
               h_name=c('儿童牛奶','养生牛奶','新出产品','其他产品','牛奶饮料','主推产品'),
               v_name=c('热销产品','普通产品','滞销产品'),
               colors = c('#E41A1C','#377EB8','#4DAF4A','#984EA3',
                          '#FF7FF0','#FFD92F'),xaxis_name='百分比',
               title='电商乳制品产品线分类统计',paper_bgcolor='#ccece6',
               plot_bgcolor='#ccece6')
}
\seealso{
\code{\link[plotly]{add_trace}},\code{\link[plotly]{layout}}
}
