% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_missing.R
\name{impute_missing}
\alias{impute_missing}
\title{Imputing Missing Value by MICE}
\usage{
impute_missing(dataframe,ord.col,ignore.predictor=NA,ignore.imputation=NA)
}
\arguments{
\item{dataframe:}{a dataframe-like object}

\item{ord.col:}{a vector of strings. It tells the mice function which columns of the
original dataframe should be treated as ordinal categorical variable.}

\item{ignore.predictor:}{a logical indicating which columns should be removed when using
mice. Default to NA. See "Details" for more information.}

\item{ignore.imputation:}{a logical indicating which columns with missing values should not
be imputed. Default to NA. See "Details" for more information.}
}
\value{
A list, with the first dataframe which is the original dataframe with
missing value been imputed by mice function and customer specified conditions,
while the second element is a  S3 object of class \link{mids} generated by mice.
}
\description{
This function will impute missing values in a dataframe by mice (stand for multivariate
imputation by chained equations). It's a handy approach of the more complicated mice
package which preserves its main feature.
}
\details{
The default imputation method in mice will be used to the dataframe, that is 'pmm' for
numerical variable, 'logreg' for categorical variable with 2 levels, 'polyreg' for
categorical variable with more than 2 levels and 'polr' for ordered categorical variables
with more than 2 levels. Detailed meaning of each methods can be found in \code{\link{mice}}.
The default setting of mice can not distinguish ordinal and nominal categorical variables,
so parameter ord.col is added to make it clear. ignore.predictor speciyies the features
that you don't want to be used when runs mice, like the id of each sample. ignore.imputation
are the features that will skip imputation. These are the features that you believe for
some reasons that should not be imputed. This function is a naive version of mice. In mice,
a standard workflow is impute a bunch of dataframe from original one and then using pooling
method for analysis. There are also many different settings to make it fit case by case.
However, such a workflow may be to complicated for basic data analysis, especially when
users have little experience. This function is designed to help. It randomly select one
imputed dataframe as the major output, so that the user can proceed to the next step of their
analysis just as using a regular dataframe, while it also keeps the whole S3 object generated
by mice as another part of the output (for the convanice of check whether imputation make sense).
For those more sophisticated users and user with special need in dealing
with missing value, they may refer to the original mice package.
}
\examples{
data("tmall_milk_sales")
test_df=tmall_milk_sales
test_df[sample(1:nrow(test_df),200),'promotion'] <- NA
test_df[sample(1:nrow(test_df),50),'feature'] <- NA
test_df[sample(1:nrow(test_df),20),'units'] <- NA
test_df[sample(1:nrow(test_df),5),'unit_price'] <- NA
test_df[sample(1:nrow(test_df),20),'label'] <- NA
## try two imputings
df_imputed1=impute_missing(test_df,ord.col="label")
sapply(df_imputed1$impute,function(x) sum(is.na(x)))
df_imputed2=impute_missing(test_df,ord.col="label",ignore.imputation = "unit_price")
sapply(df_imputed2$impute,function(x) sum(is.na(x)))
## to check whether the imputation make sense
densityplot(df_imputed1$pool,scales=list(x=list(relation='free')))
}
\references{
Stef van Buuren and Karin Groothuis-Oudshoorn(2011) mice: Multivariate
Imputation by Chained Equations in R. \emph{Journal of Statistical Software, Articles}
Volumn 45, Page 1-67, DOI: 10.18637/jss.v045.i03
Michy Alice, Imputing Missing Data with R: mice Package.
\url{https://datascienceplus.com/imputing-missing-data-with-r-mice-package/}
Klodian Dhana, Handling missing data with mice package: a simple approach.
\url{https://datascienceplus.com/handling-missing-data-with-mice-package-a-simple-approach/}
}
\seealso{
\code{\link{mice}}
}
