% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_predict.R
\name{lin_predict}
\alias{lin_predict}
\title{Linear Extrapolation Using Dynamic Linear Model}
\usage{
lin_predict(dataframe,ts.idx,t_ahead,addCI,ctg.idx,extra_names,xaxis_name,
                   yaxis_name,title,...)
}
\arguments{
\item{...}{other plot settings. See \code{\link[plotly]{layout}} for more information}

\item{dataframe:}{a dataframe object}

\item{ts.idx:}{a character or a numeric, indicate the column in the dataframe that you
want to get prediction with}

\item{t_ahead:}{an integer, how many predictions whould be made}

\item{addCI:}{a logical, whether to add confidence interval to the visualization}

\item{ctg.idx:}{a character or a numeric, indicate the column in the dataframe that will
be used as the x axis index in the plot. See "Details" for more information}

\item{extra_names:}{a vector of strings, will be used as the x axis tick text for
prediction points. See "Details" for more information}

\item{xaxis_name:}{the name of x axis for the plot}

\item{yaxis_name:}{the name of y axis for the plot}

\item{title:}{the name of the plot}
}
\value{
A list with two elements called pred.mtx and pred.plot respectively.
The pred.mtx is a dataframe containing the predict mean, 95% confidence interval for
the series and period you specified. The pred.plot is a plotly object which is a
visualization of how the dynamic linear model with trend fits original data and
the trend of future prediction and confidence intervals.
}
\description{
\code{lin_predict} generate extrapolation of a numeric variable based on dynamic linear
model. The result is shown as a dataframe containing predict mean and confidence interval
as well as a plot for visualization
}
\details{
It's really often to do prediction or extrapolation of a numerical variable, especially
when it is a time series. This function accomplish this task by using dynamic linear
model, more specifically, random walk plus trend model. The explicit formula of the
model can be found in Petris' book Dynamic Linear Model with R. The procedure is also
suggested in the book. The function will first build a random walk plus trend model
using \code{\link[dlm]{dlm}}, then estiamte parameters in dlm by \code{\link[dlm]{dlmMLE}}.
After that, \code{\link[dlm]{dlmFilter}} will be used to filter and get one-step-ahead
forecast for model checking. Finally using \code{\link[dlm]{dlmForecast}}, one can get
the forecast. The plot is a nice way to see whether the model works, by judging the gap
between model fitting and real data. When setting x axis of the plot, if 'ctg.idx' in
the function is specified, 'extra_names' should also be specified based on the index
choosen by 'ctg.idx', or just miss both of them, then a sequence start from 1 with length
equal to the number of real data plus extrapolations will be applied as tick text of x
axis. Only given value to either one of them will generate error. Note that since the
extrapolation is done using a random walk plus trend model, you must have some
confidence that the series you want to forecast do have some trend as a priori,
(it is the case for many macroeconmic series), otherwise the model may not converge
and the result can be badly wrong or the prediciton may have really large variance that
results in the confidence interval making no sense.
}
\examples{
data("macro_data_chn")
GDP_predict=lin_predict(dataframe=macro_data_chn,ts.idx='GDP',t_ahead=3,
                        addCI=T,xaxis_name='Time',yaxis_name='GDP(元)',
                        title='GDP未来三年预测')
GDP_predict$pred.mtx
GDP_predict$pred.plot
GDP_predict=lin_predict(dataframe=macro_data_chn,ts.idx='GDP',t_ahead=3,addCI=T,
                        ctg.idx=1,extra_names = c(2017,2018,2019),
                        xaxis_name='年份',yaxis_name='GDP(元)',title='GDP未来三年预测',
                        legend=list(x=0.72,y=0.1,bgcolor='transparent'),
                        margin=list(t=45,l=45,r=18),paper_bgcolor='#ccece6')
GDP_predict$pred.mtx
GDP_predict$pred.plot
}
\references{
Petris, Giovanni and Petrone, Sonia and Campagnoli, Patrizia(2009)
Dynamic Linear Models with R. \emph{Journal of Applied Statistics}
Volumn 38, pages 31-84, DOI: 10.1007/b135794_2
Petris, Giovanni and Petrone, Sonia and Campagnoli, Patrizia(2009)
Dynamic Linear Models with R. Book published by Springer.
ISBN 978-0-387-77237-0
Giovanni Petris, dlm: an R package for Bayesian analysis of Dynamic Linear Models
Vignette of R package dlm. Available at:
\url{https://cran.r-project.org/web/packages/dlm/vignettes/dlm.pdf}
}
\seealso{
\code{\link[dlm]{dlm}},\code{\link[dlm]{dlmMLE}}
\code{\link[dlm]{dlmFilter}},\code{\link[dlm]{dlmForecast}}
}
\author{
Jizhou Kang <jizhou_kang@hotmail.com>
}
