% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_ann_plot.R
\name{line_ann_plot}
\alias{line_ann_plot}
\title{Line Plots with Information about Turning Points}
\usage{
line_ann_plot(dataframe,ctg.idx,num.idx,condition,condition.idx,colors,
                     events, marker_pos_x,ann_pos_x,text_pos_x,ann_pos_y,
                     marker_refer,marker_pos_adj,yaxis_name,title,
                     marker_color='rgb(246,78,139)',...)
}
\arguments{
\item{...}{other parameters for plotting, see "Details" and \code{\link[plotly]{layout}}
for more information}

\item{dataframe:}{a dataframe object}

\item{ctg.idx:}{a character or an integer, indicate which column will be selected as the
variable to show in x axis}

\item{num.idx:}{a character or an integer, indicate which column will be selected as the
variable to show in y axis}

\item{condition:}{a vector, indicate how to select part of the dataframe}

\item{condition.idx:}{a character or an integer, indicate which column is the condition in.
Should be given some value if the condition appears in multiple columns in the dataframe or
condition is missing}

\item{colors:}{a vector, the i-th element represent the color that will be used to draw
the i-th line, which refers to the numeric value of i-th condition.}

\item{events:}{a vector of strings, explanation of each turning point}

\item{marker_pos_x:}{a vector, should be the same length as events, indicate the x axis
position of turning points markers}

\item{ann_pos_x:}{a numeric, indicate the x axis position of markers in annotation}

\item{text_pos_x:}{a numeric, indicate the x axis position of explanations in annotation}

\item{ann_pos_y:}{a vector of numeric, indicate the y axis position of markers in annotation}

\item{marker_refer:}{which line should be referred to when putting turning points markers}

\item{marker_pos_adj:}{should the marker in the plot have some position adjusting in y axis}

\item{yaxis_name:}{name of y axis}

\item{title:}{the title of the plot}

\item{marker_color:}{the color of marker, default to 'rgb(246,78,139)'(alike purple)}
}
\value{
This function will return an object of class plotly. By calling the function you
will get a plot which is a series plot with markers and annotations marking turning points.
}
\description{
\code{line_ann_plot} will add user specified turning points and corresponding explanations
to a regular line plots.
}
\details{
This function uses the package \code{\link{plotly}} to make fancy line plots with annotations
of the turning points of the dataframe. To make the plot, you will need a dataframe
containing at least three columns, the column for x axis and y axis of the plot, as well
as a categorical column which will be used to select data or display different lines.
The function already have default settings for plot which makes it handy to users with
little experience in using plotly. For those advanced users, more plotting parameters can
be used because of the three dots structure. Such options may include: xaxis, legend,
and paper_bgcolor. See \code{\link[plotly]{layout}} for more information.
}
\examples{
data(dairy_fin_charts)
line_ann_plot(dataframe=dairy_fin_charts,ctg.idx='Year',num.idx='profit',
              condition=c('伊利股份','蒙牛股份','光明乳业'),
              colors = c("#00526d","#de6e6e","rgb(50,171,96)"),
              events = c("中国奶制品污染事件&金融危机","公布和实施经济刺激计划",
                         "“互联网+”：传统行业进入电商时代","中央一号文件：全面振兴奶业"),
              marker_pos_x = c(2008,2009,2012,2017),
              ann_pos_x = 2006, text_pos_x = 0.1, ann_pos_y = c(78,74,70,66),
              marker_refer = c('光明乳业','蒙牛股份','伊利股份','伊利股份'),
              yaxis_name = '利润总额（亿元）',title = '乳制品企业利润变动与行业重要事件')
## customize some advanced parameters
line_ann_plot(dataframe=dairy_fin_charts,ctg.idx='Year',num.idx='profit',
              condition=c('伊利股份','蒙牛股份','光明乳业'),
              colors = c("#00526d","#de6e6e","rgb(50,171,96)"),
              events = c("中国奶制品污染事件&金融危机","公布和实施经济刺激计划",
                         "“互联网+”：传统行业进入电商时代","中央一号文件：全面振兴奶业"),
              marker_pos_x = c(2008,2009,2012,2017),
              ann_pos_x = 2006, text_pos_x = 0.1, ann_pos_y = c(78,74,70,66),
              marker_refer = c('光明乳业','蒙牛股份','伊利股份','伊利股份'),
              marker_pos_adj=3,
              yaxis_name = '利润总额（亿元）',
              title = '乳制品企业利润变动与行业重要事件',
              legend = list(x=0.5,y=0.1,orientation='h',
                            font=list(size=10),bgcolor='transparent'),
              xaxis = list(showgrid=T,nticks=12,ticks="outside",title="年份"),
              paper_bgcolor='#ccece6')
}
\references{
Time Series Charts By the Economist in R Using Plotly
Published by Roddhiman in Dashboards, Data Visualization, R
\url{https://moderndata.plot.ly/time-series-charts-by-the-economist-in-r-using-plotly/}
}
\seealso{
\code{\link[plotly]{add_trace}},\code{\link[plotly]{layout}}
}
\author{
Jizhou Kang <jizhou_kang@hotmail.com>
}
