% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_plot.R
\name{lines_plot}
\alias{lines_plot}
\title{Making Lines Plot with Multiple Options}
\usage{
lines_plot(dataframe,ctg.idx,num.idx,condition,condition.idx,
                  colors,mode='lines',yaxis_name="",linewidth=2,
                  title,...)
}
\arguments{
\item{...}{other parameters for plotting, see "Details" and \code{\link[plotly]{layout}}
for more information}

\item{dataframe:}{a dataframe object}

\item{ctg.idx:}{a character or an integer, indicate which column will be selected as the
variable to show in x axis}

\item{num.idx:}{a character or an integer, indicate which column(s) will be selected as the
variable to show in y axis}

\item{condition:}{a vector, indicate how to select part of the dataframe}

\item{condition.idx:}{a character or an integer, indicate which column is the condition in.
Should be given some value if the condition appears in multiple columns in the dataframe}

\item{colors:}{a vector, the i-th element represent the color that will be used to draw
the i-th line, which refers to the numeric value of i-th condition}

\item{mode:}{how should each line displayed, usually be set as 'lines' or 'lines+markers',
see \url{https://plot.ly/r/reference/#scatter} for more information}

\item{yaxis_name:}{name of y axis}

\item{linewidth:}{the width of each line, default to 2}

\item{title:}{the title of the plot}
}
\value{
This function will return an object of class plotly. By calling the function you
will get a lines plot which is displayed as you sepcified in the function.
}
\description{
Line plots are the most basic and commonly-used plots in data visualization.
\code{lines_plot} makes line plot with multiple choice on the display.
}
\details{
This function makes it easier to draw line plot no matter which format your data have.
For beginners, it takes relatively little arguments and provided a fancy look line plot.
For sophisticated users, flexibility can be achieved by the three dots structure.
Some unlisted option can be take into the function, such as legend, paper_bgcolor and
margin. See \code{\link[plotly]{layout}} for more information.
}
\examples{
data("dairy_fin_charts")
lines_plot(dataframe=dairy_fin_charts,ctg.idx = 5,num.idx = 2,
           condition=c('伊利股份','蒙牛股份','光明乳业'),
           colors=c("#00526d","#de6e6e","#32ab60"),
          yaxis_name = '营业收入',linewidth = 4,
           title='乳制品企业营业收入图')
lines_plot(dataframe=dairy_fin_charts,ctg.idx = 5,num.idx = 2,
           condition=c('伊利股份','蒙牛股份','光明乳业'),
           colors=c("#00526d","#de6e6e","#32ab60"),
           mode='lines+markers',linewidth = 2,
          title='乳制品企业营业收入图',
          yaxis_name='营业收入',
          xaxis=list(showgrid=F,nticks=10,ticklen=4,tickangle=-45,
                     ticks='outside',tickmode="array",
                     type='category',title="年份"),
          yaxis=list(visible=F),
          legend=list(x=0.5,y=0.1,orientation='h',
                      font=list(size=10),bgcolor="transparent"),
          paper_bgcolor='#ccece6',margin=list(t=32,l=32,r=32))
}
\seealso{
\code{\link[plotly]{plot_ly}},\code{\link[plotly]{layout}}
}
