% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_split.R
\name{lines_split}
\alias{lines_split}
\title{Plot Lines with Different Range}
\usage{
lines_split(dataframe,ctg.idx,num.idx,condition,condition.idx,
                   colors,mode='lines',yaxis_name="",linewidth=2,
                   title,...)
}
\arguments{
\item{...}{other parameters for plotting, see "Details" and \code{\link[plotly]{layout}}
for more information}

\item{dataframe:}{a dataframe object}

\item{ctg.idx:}{a character or an integer, indicate which column will be selected as the
variable to show in x axis}

\item{num.idx:}{a vector of character or an integer, indicate which column(s) will
be selected as the variable to show in y axis}

\item{condition:}{a vector, indicate how to select part of the dataframe}

\item{condition.idx:}{a character or an integer, indicate which column is the condition in.
Should be given some value if the condition appears in multiple columns in the dataframe}

\item{colors:}{a vector, the i-th element represent the color that will be used to draw
the i-th line, which refers to the numeric value of i-th condition.}

\item{mode:}{how should each line displayed, usually be set as 'lines' or 'lines+markers',
see \url{https://plot.ly/r/reference/#scatter} for more information}

\item{yaxis_name:}{name of y axis}

\item{linewidth:}{the width of each line, default to 2}

\item{title:}{the title of the plot}
}
\description{
\code{lines_split} uses subplot to plot numeric variables with various range
but share same index.
}
\details{
When plotting different variables that shares same index but have different range, using
one plot may generate big problem since some of the variable may just be a flat line. To
avoid this problem, one possible solution is to use subplot. This function is designed for
this purpose. The three dots structure inside the inner function lines_split are
options for the general settings of layout, see \code{\link[plotly]{layout}} for
more information.
}
\examples{
data("macro_data_chn")
lines_split(dataframe=macro_data_chn,ctg.idx = 'year',
            num.idx = c(3,5,6,9),
            colors=c("#00526d","#de6e6e","#32ab60","#ff8000"),
            title="一些宏观经济指标走势",
            xaxis=list(showgrid=F,ticklen=4,nticks=3,title="年份"),
            legend=list(x=0.5,y=1.05,orientation='h',bgcolor='transparent'),
            paper_bgcolor='#ccece6',margin=list(t=32,l=32,r=32))
}
\seealso{
\code{\link[plotly]{plot_ly}}, \code{\link[plotly]{layout}},
\code{\link{lines_split_plot}}
}
