% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinery_fin_charts.R
\docType{data}
\name{machinery_fin_charts}
\alias{machinery_fin_charts}
\title{Financial Charts of Chinese Machinery Manufacturing Companies(2000-2017)}
\format{A dataframe containing 5457 rows and 49 columns}
\usage{
data(machinery_fin_charts)
}
\description{
Balance sheet and characteristic data of Chinese machinery manufacturing
companies whose stock is traded on Chinese stock market
from 2000 to 2017.
}
\references{
Wind database of Chinese stock market data
}
\keyword{datasets}
