% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_data_chn.R
\docType{data}
\name{macro_data_chn}
\alias{macro_data_chn}
\title{Macroeconomic Data of China(2008-2016)}
\format{A dataframe containing 9 rows and 12 columns}
\usage{
data(macro_data_chn)
}
\description{
Macroeconomic data of China such as CPI, PPI, GDP and also dairy industry data,
such as indusrtial ROE and price of related goods, from 2007 to 2016.
}
\references{
Wind database of Chinese industry and macroeconomic data
}
\keyword{datasets}
