% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miss_prep.R
\name{miss_prep}
\alias{miss_prep}
\title{Missing Value Pattern Checking and Processing}
\usage{
miss_prep(dataframe,remove.column=TRUE,remove.row=FALSE)
}
\arguments{
\item{dataframe:}{a dataframe-like object}

\item{drop.column:}{a logical indicating whether to automatically drop columns.
Default to TRUE. See "Details" for more information}

\item{drop.row:}{a logical indicating whether to automatically drop rows.
Default to FALSE. See "Details" for more information}
}
\value{
Two dataframes, the dataframe called "pattern" contains the pattern of missings in
the original dataframe, while the dataframe called "df" contains the processed dataframe,
which maybe used for further processing such as imputation.
}
\description{
This function will detect the pattern of missing values and output the pattern as
a dataframe. It can also process the dataframe by eliminating the missed columns or
rows when needed.
}
\details{
The output dataframe "pattern" gives you a clear image about the pattern of missing value
in your dataframe when considering combinations of features. The output dataframe "df" is
the processed dataframe. By default, it will automatically drop columns with more than 30%
missing while keep rows. This is just some usual way suggested by data scientist community.
You can change the remove.row to TRUE to let the function helps you drop samples with
more than 50% features missing, or set remove.column to FALSE to ignore drop columns.
}
\examples{
## Generate some NAs in the dataframe for illustration
data("tmall_milk_sales")
test_df=tmall_milk_sales
test_df[sample(1:nrow(test_df),200),'promotion'] <- NA
test_df[sample(1:nrow(test_df),50),'feature'] <- NA
test_df[sample(1:nrow(test_df),20),'units'] <- NA
test_df[sample(1:nrow(test_df),5),'unit_price'] <- NA
## test function
miss=miss_prep(test_df)
miss_pattern=miss$pattern
df_processed=miss$df
}
\references{
Michy Alice, Imputing Missing Data with R: mice Package.
\url{https://datascienceplus.com/imputing-missing-data-with-r-mice-package/}
}
\seealso{
\code{\link{aggr}}
}
