% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom_ctg2num.R
\name{nom_ctg2num}
\alias{nom_ctg2num}
\title{Transform nominal categorical variable into numerical variable}
\usage{
nom_ctg2num(dataframe,col.id,col.name=NA,drop)
}
\arguments{
\item{dataframe:}{a dataframe-like object}

\item{col.id:}{a vector of integers, specify which column in the dataframe
that you need to transfer}

\item{col.name:}{a vector of characters, another way to specify
the to-be-transformed column, will be disregarded if col.id is specified}

\item{drop:}{tells the function which dummied column you want to drop. The dropped one
will be served as the baseline in like regression model. Can be missed.}
}
\value{
a dataframe of dummy variables generated by choosen columns from original dataframe
}
\description{
Transform a nominal categorical column into multiple numerical columns with dummies,
can based on customer specified rule.
}
\details{
The function will match each element in a categorical variable with a numeric value
by the mapping rule specified in the function input.
}
\examples{
data(tmall_milk_sales)
test_df=tmall_milk_sales
nom_ctg2num(test_df,col.id=c(1,2,3,4,5))
nom_ctg2num(test_df,col.name=c('label','promotion'),drop=c('H','N'))
\dontrun{
  # be careful when having NAs, it will be treated as a new dummy
  test_df$label[12:15]=NA
  nom_ctg2num(test_df,col.name='label',drop='H')
}
}
\seealso{
\code{\link[dummies]{dummy}}
}
