% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_ctg2num.R
\name{ord_ctg2num}
\alias{ord_ctg2num}
\title{Transform ordinal categorical variable into numerical variable}
\usage{
ord_ctg2num(dataframe,col.id,col.name=NA,permutation,
numeric_levels=seq(1,length(permutation),1))
}
\arguments{
\item{dataframe:}{a dataframe-like object}

\item{col.id:}{an integer, specify which column in the dataframe that you need to transfer}

\item{col.name:}{a character, another way to specify the to-be-transformed column,
will be disregarded if col.id is specified}

\item{permutation:}{a vector of string which is a permutation of unique element in
your categorical variable. It will serve as the rule for assign numeric values.
The first element in this permutation will be assigned the smallest numeric value}

\item{numeric_levels:}{a numeric vector which determine the numeric values to be
assigned to the categorical variable, default to consecutive sequence of length
equal to permutation and start from 1}
}
\value{
Numerical vairable which is transformed from the assigned categorical variable
based on the specified order.
}
\description{
Transform a ordinal categorical column into a numerical column, based on
customer specified rule.
}
\details{
The function will match each element in a categorical variable with a numeric value
by the mapping rule specified in the function input.
}
\examples{
data(tmall_milk_sales)
test_df=tmall_milk_sales
test_df$label[12:15]=NA
ord_ctg2num(test_df,col.name = 'label',permutation = c('P','O','H'))
ord_ctg2num(test_df,col.id = 11,permutation = c('P','O','H'))
ord_ctg2num(test_df,col.id = 11,permutation = c('P','O','H'),
numeric_levels=c(1,5,10))
\dontrun{
  # The following code will generate error,
  # because permutation not match
  ord_ctg2num(test_df,col.id = 11,permutation = c('p','o','h'))
}
}
