% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm_basic.R
\name{plm_basic}
\alias{plm_basic}
\title{Stepwise Numerical Variable Selection for Panel Linear Regression}
\usage{
plm_basic(dataframe,id.idx,t.idx,dep.idx,control.idx,num.idx,step)
}
\arguments{
\item{dataframe:}{a dataframe object}

\item{id.idx:}{a string or an integer, indicate which column of the dataframe will be used
as individual index in panel dataframe}

\item{t.idx:}{a string or an integer, indicate which column of the dataframe will be used
as time index in panel dataframe}

\item{dep.idx:}{a string or an integer, indicate which column of the dataframe will be used
as the dependent variable, i.e. y, in the regression function}

\item{control.idx:}{a vector of strings or integers, but not both, indicate which column(s)
of the dataframe will be used as control variables in regression function}

\item{num.idx:}{a vector of strings or integers, but not both, indicate which column(s) of
the dataframe will be used as independent variables of interest in regression function.
See "Details" for more information}

\item{step:}{how many forward steps should be made. See "Details" for more information}
}
\value{
A table that shows the summary information of all selected panel linear regression
models as well as a list containing all this models for further exploration.
}
\description{
\code{plm_basic} applied forward stepwise variable selection for panel data linear
regression. It can serve as a baseline for further model selection
}
\details{
The function works like this: it start with the regression formula that contains only
all control variables. A model selection function will be applied
to determine which model among the three ('OLS','fixed effect','random effect') shall
be used. Then at each step, all possible variables, start with variables indicated by
"num.idx", will be add to the fourmula one at a time. For each time, the same  model
selection function will be applied to determine which model should be used, and the
p-value of the new variable adding to the formula in the selected model will be record.
Then for all the possible variables in this step, the one with smallest p-value, i.e. the
one which is the most significant variable among all possible choice will be add to
the formula permentately. It will then be eliminated from possible options that may
be add to the formula, and the corresponding model will be record for output. This loop
will be ran "step" times and a list containing "step"+1 models will be output as a list,
together with a summary table generated by \code{\link{stargazer}}. This is the standard
way for doing explorational panel regression. Note that since
within model will eliminate time invariant variables, the 'num.idx' should only contain
those variables that varied by time. Meanwhile, since it's a stepwise selection method,
dummy variable generated by nominal categorical variables with more than two levels should
also be avoided in "num.idx". For the summary table generated by stargazer looks clearly
in RStudio, we suggest steps less than or equal to 6, although one can have larger "step"
up to the length of "num.idx"
}
\examples{
data("estate_fin_charts")
estate_fin_charts$lg_asset=log(estate_fin_charts$asset)
model.list=plm_basic(dataframe=estate_fin_charts,id.idx=3,t.idx='Year',dep.idx='lg_income',
                     control.idx=c('mkt_price','lg_asset'),
                     num.idx=c(5,6,7,8,10,11,14,15),step=5)
}
\references{
Jeffrey M. Wooldridge Introductory Econometrics: A Modern Approach,
Fifth edition
James Lee Ray, What Should Be Controlled for?(2003),
 \emph{Conflict Management and Peace Science}, Volume 20;1
 DOI: 10.1177/07388942032000201
}
\seealso{
\code{\link[plm]{plm}}
}
\author{
Jizhou Kang \email{jizhou_kang@hotmail.com}
}
