% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_charts.R
\name{polar_charts}
\alias{polar_charts}
\title{Drawing Radar Plot}
\usage{
polar_charts(dataframe,colors,fills,fillcolors,title,...)
}
\arguments{
\item{...}{other plot parameters. See \code{\link[plotly]{layout}} for more information}

\item{dataframe:}{a dataframe object}

\item{colors:}{a vector, the palette used to draw lines}

\item{fills:}{a categorical string of either 'none' or 'toself', whether to fill the polygon
created by the line}

\item{fillcolors:}{a vector, the palette used to fill polygons, will be ignored if related
'fills' is set to 'none'}

\item{title:}{the name of the plot}
}
\value{
This function will return an object of class plotly. By calling the function you
will get a polar plot for your dataframe.
}
\description{
\code{polar_charts} provides polar charts of the dataframe, each row will be a trace in
the polar charts.
}
\details{
The row name will be served as the name of each trace that will also be shown in the legend.
The column name of the dataframe will be used as the tick text of the angular axis of
the polar charts. Note that if choose 'fills' as 'toself', then the relative 'fillcolors'
should be somewhat transparent, which can be done by setting colors using 'rgba'.
}
\examples{
## prepare data, it's generate from analysis of NASDAQ: MDLZ
MDLZ=c(6,5,2,6,1,5,1,1,3,1,2,2,10,10,9,9)
others=c(7,9,6,6,3,2,2,2,3,2,4,4,10,10,9,9)
polar.dataframe=rbind(MDLZ,others)
row.names(polar.dataframe)=c('亿滋国际','市场同业竞争者均值')
colnames(polar.dataframe)=c('股息收益','净资产收益率','资产回报率','息税前利润',
                            '销售增长率','净收入增长率','营收增长期望','每股盈余期望',
                            '市盈率','市售率','企业价值倍数','市现率',
                            '总市值','成交量','波动性','风险系数')
## a simple example
polar_charts(dataframe=polar.dataframe,colors=c('#FF7F00','#33A02C'),fills=c('toself','none'),
             fillcolors=c('#FDBF6F','#B2DF8A'),title='亿滋国际和同行竞争对手股票指标对比',
             margin=list(t=56))
## another way
polar_charts(dataframe=polar.dataframe,colors=c('#FF7F00','#33A02C'),fills=c('toself','toself'),
             fillcolors=c('rgba(253,191,111,0.3)','rgba(178,223,138,0.3)'),
             title='亿滋国际和同行竞争对手股票指标对比',
             legend=list(x=1,y=0,bgcolor='transparent',font=list(size=14,family='heiti')),
             margin=list(t=56),paper_bgcolor='#ccece6')
}
\seealso{
\code{\link[plotly]{plot_ly}},\code{\link[plotly]{layout}}
}
