% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmall_milk_sales.R
\docType{data}
\name{tmall_milk_sales}
\alias{tmall_milk_sales}
\title{Tmall Market Milk Sales Data(Accumulated until 2018-07)}
\format{A dataframe containing 336 rows and 11 columns}
\usage{
data(tmall_milk_sales)
}
\description{
Top 10 most popular milk producer's product sales data on Tmall Market,
which is the largest Chinese online market. The data contains features of the milk products,
such as weight and price, as well as total sales volume accumulated until 2018-07.
}
\references{
Tmall at \link{www.tmall.com}
}
\keyword{datasets}
